Split EntitySchema interfaces into separate files
Create separate files for EntitySchemaTable, EntitySchemaColumn, EntitySchemaIndex and EntitySchemaRelation
